Improvements on the treatment of SPOT data
The changes in this commit solve some issues in the SPOT processing
chain. These are to do with new definition of the emulators, as well as
with a bug that resulted in an error due to the loop iterator assuming
that the scanned files had been stored sorted by time.